feat: show selected filters inside dropdown button + checkbox switcher
