+++ src/search/views/Search.tsx
@@ -15,7 +15,6 @@ import {
 	Select,
 	Spacer,
 	Spinner,
-	TagList,
 	TextInput,
 	Thumbnail,
 	Toolbar,
@@ -31,8 +30,6 @@ import {
 	cloneDeep,
 	compact,
 	every,
-	find,
-	flatten,
 	get,
 	isArray,
 	isEmpty,
@@ -40,14 +37,14 @@ import {
 	isNil,
 	isPlainObject,
 	pickBy,
-	remove,
 } from 'lodash-es';
 import queryString from 'query-string';
 import { RouteComponentProps } from 'react-router';
 import { Link } from 'react-router-dom';
 
-import { CheckboxDropdown } from '../../shared/components/CheckboxDropdown/CheckboxDropdown';
-import { CheckboxModal, CheckboxOption } from '../../shared/components/CheckboxModal/CheckboxModal';
+import { CheckboxDropdown } from '../../shared/components/CheckboxSwitcher/CheckboxDropdown/CheckboxDropdown';
+import { CheckboxModal } from '../../shared/components/CheckboxSwitcher/CheckboxModal/CheckboxModal';
+import { CheckboxOption } from '../../shared/components/CheckboxSwitcher/CheckboxSwitcher';
 import { DateRangeDropdown } from '../../shared/components/DateRangeDropdown/DateRangeDropdown';
 import { formatDate } from '../../shared/helpers/formatters/date';
 import { formatDuration } from '../../shared/helpers/formatters/duration';
@@ -67,12 +64,6 @@ interface SortOrder {
 interface SearchResults {
 	count: number;
 	items: Avo.Search.ResultItem[];
-}
-
-interface TagInfo {
-	label: string;
-	prop: Avo.Search.FilterProp;
-	value: any;
 }
 
 const DEFAULT_FORM_STATE: Avo.Search.Filters = {
@@ -293,7 +284,8 @@ export const Search: FunctionComponent<SearchProps> = ({ history, location }: Se
 					label = languageCodeToLabel(option.option_name);
 				}
 				return {
-					label: `${label} (${option.option_count})`,
+					label,
+					optionCount: option.option_count,
 					id: option.option_name,
 					checked: ((formState[propertyName] as string[]) || []).includes(option.option_name),
 				};
@@ -324,7 +316,8 @@ export const Search: FunctionComponent<SearchProps> = ({ history, location }: Se
 			(option: Avo.Search.OptionProp): CheckboxOption => {
 				const label = capitalize(option.option_name);
 				return {
-					label: `${label} (${option.option_count})`,
+					label,
+					optionCount: option.option_count,
 					id: option.option_name,
 					checked: ((formState[propertyName] as string[]) || []).includes(option.option_name),
 				};
@@ -382,119 +375,119 @@ export const Search: FunctionComponent<SearchProps> = ({ history, location }: Se
 			</div>
 		);
 	};
+	//
+	// const getTagInfos = (filterProp: Avo.Search.FilterProp, filterValue: any): TagInfo[] => {
+	// 	// Do not render query filter or empty filters
+	// 	if (
+	// 		filterProp === 'query' ||
+	// 		filterValue === '' ||
+	// 		filterValue === [] ||
+	// 		(isArray(filterValue) && every(filterValue, filter => !filter)) // Array of empty strings
+	// 	) {
+	// 		return [];
+	// 	}
+	//
+	// 	// Render date range option filters
+	// 	if (isPlainObject(filterValue)) {
+	// 		if (filterValue.gte && filterValue.lte) {
+	// 			return [
+	// 				{
+	// 					label: `${formatDate(filterValue.gte)} - ${formatDate(filterValue.lte)}`,
+	// 					prop: filterProp,
+	// 					id: filterValue,
+	// 				},
+	// 			];
+	// 		}
+	// 		if (filterValue.gte) {
+	// 			return [
+	// 				{
+	// 					label: `na ${formatDate(filterValue.gte)}`,
+	// 					prop: filterProp,
+	// 					id: filterValue,
+	// 				},
+	// 			];
+	// 		}
+	// 		if (filterValue.lte) {
+	// 			return [
+	// 				{
+	// 					label: `voor ${formatDate(filterValue.lte)}`,
+	// 					prop: filterProp,
+	// 					id: filterValue,
+	// 				},
+	// 			];
+	// 		}
+	// 		return []; // Do not render a filter if date object is empty: {gte: "", lte: ""}
+	// 	}
+	//
+	// 	// Render multi option filters
+	// 	if (isArray(filterValue)) {
+	// 		return filterValue.map((filterVal: string) => {
+	// 			let label = filterVal;
+	// 			if (filterProp === 'language') {
+	// 				label = languageCodeToLabel(filterVal);
+	// 			}
+	// 			return {
+	// 				label,
+	// 				prop: filterProp,
+	// 				id: filterVal,
+	// 			};
+	// 		});
+	// 	}
+	//
+	// 	console.error('Failed to render selected filter: ', filterProp, filterValue);
+	// 	return [];
+	// };
 
-	const getTagInfos = (filterProp: Avo.Search.FilterProp, filterValue: any): TagInfo[] => {
-		// Do not render query filter or empty filters
-		if (
-			filterProp === 'query' ||
-			filterValue === '' ||
-			filterValue === [] ||
-			(isArray(filterValue) && every(filterValue, filter => !filter)) // Array of empty strings
-		) {
-			return [];
-		}
-
-		// Render date range option filters
-		if (isPlainObject(filterValue)) {
-			if (filterValue.gte && filterValue.lte) {
-				return [
-					{
-						label: `${formatDate(filterValue.gte)} - ${formatDate(filterValue.lte)}`,
-						prop: filterProp,
-						value: filterValue,
-					},
-				];
-			}
-			if (filterValue.gte) {
-				return [
-					{
-						label: `na ${formatDate(filterValue.gte)}`,
-						prop: filterProp,
-						value: filterValue,
-					},
-				];
-			}
-			if (filterValue.lte) {
-				return [
-					{
-						label: `voor ${formatDate(filterValue.lte)}`,
-						prop: filterProp,
-						value: filterValue,
-					},
-				];
-			}
-			return []; // Do not render a filter if date object is empty: {gte: "", lte: ""}
-		}
-
-		// Render multi option filters
-		if (isArray(filterValue)) {
-			return filterValue.map((filterVal: string) => {
-				let label = filterVal;
-				if (filterProp === 'language') {
-					label = languageCodeToLabel(filterVal);
-				}
-				return {
-					label,
-					prop: filterProp,
-					value: filterVal,
-				};
-			});
-		}
-
-		console.error('Failed to render selected filter: ', filterProp, filterValue);
-		return [];
-	};
-
-	const renderSelectedFilters = () => {
-		const tagInfos: TagInfo[] = flatten(
-			(Object.keys(formState) as Avo.Search.FilterProp[]).map((filterProp: Avo.Search.FilterProp) =>
-				getTagInfos(filterProp, formState[filterProp])
-			)
-		);
-		const tagLabels = tagInfos.map((tagInfo: TagInfo) => tagInfo.label);
-		if (tagLabels.length > 1) {
-			tagLabels.push('Alle filters wissen');
-		}
-		return (
-			<Spacer margin="bottom-large">
-				<TagList
-					closable={true}
-					swatches={false}
-					onTagClosed={async (tagLabel: string) => {
-						if (tagLabel === 'Alle filters wissen') {
-							deleteAllFilters();
-						} else {
-							const tagInfo = find(tagInfos, (tagInfo: TagInfo) => tagInfo.label === tagLabel);
-							if (tagInfo) {
-								await deleteFilter(tagInfo);
-							}
-						}
-					}}
-					tags={tagLabels}
-				/>
-			</Spacer>
-		);
-	};
-
-	const deleteFilter = async (tagInfo: TagInfo) => {
-		if (isPlainObject(tagInfo.value) && (tagInfo.value.gte || tagInfo.value.lte)) {
-			setFormState({
-				...formState,
-				[tagInfo.prop]: DEFAULT_FORM_STATE[tagInfo.prop],
-			});
-			return;
-		}
-		if (isArray(formState[tagInfo.prop])) {
-			const filterArray: string[] = formState[tagInfo.prop] as string[];
-			remove(filterArray, filterItem => filterItem === tagInfo.value);
-			setFormState({
-				...formState,
-				[tagInfo.prop]: filterArray,
-			});
-		} else {
-			console.error('Failed to remove selected filter: ', tagInfo.prop, tagInfo.value);
-		}
-	};
+	// const renderSelectedFilters = () => {
+	// 	const tagInfos: TagInfo[] = flatten(
+	// 		(Object.keys(formState) as Avo.Search.FilterProp[]).map((filterProp: Avo.Search.FilterProp) =>
+	// 			getTagInfos(filterProp, formState[filterProp])
+	// 		)
+	// 	);
+	// 	const tags = tagInfos.map((tagInfo: TagInfo) => ({ label: tagInfo.label, id: tagInfo.id }));
+	// 	// if (tags.length > 1) {
+	// 	// 	tags.push('Alle filters wissen');
+	// 	// }
+	// 	return (
+	// 		<Spacer margin="bottom-large">
+	// 			<TagList
+	// 				closable={true}
+	// 				swatches={false}
+	// 				onTagClosed={async (tagLabel: string) => {
+	// 					if (tagLabel === 'Alle filters wissen') {
+	// 						deleteAllFilters();
+	// 					} else {
+	// 						const tagInfo = find(tagInfos, (tagInfo: TagInfo) => tagInfo.label === tagLabel);
+	// 						if (tagInfo) {
+	// 							await deleteFilter(tagInfo);
+	// 						}
+	// 					}
+	// 				}}
+	// 				tags={tags}
+	// 			/>
+	// 		</Spacer>
+	// 	);
+	// };
+	//
+	// const deleteFilter = async (tagInfo: TagInfo) => {
+	// 	if (isPlainObject(tagInfo.id) && (tagInfo.id.gte || tagInfo.id.lte)) {
+	// 		setFormState({
+	// 			...formState,
+	// 			[tagInfo.prop]: DEFAULT_FORM_STATE[tagInfo.prop],
+	// 		});
+	// 		return;
+	// 	}
+	// 	if (isArray(formState[tagInfo.prop])) {
+	// 		const filterArray: string[] = formState[tagInfo.prop] as string[];
+	// 		remove(filterArray, filterItem => filterItem === tagInfo.id);
+	// 		setFormState({
+	// 			...formState,
+	// 			[tagInfo.prop]: filterArray,
+	// 		});
+	// 	} else {
+	// 		console.error('Failed to remove selected filter: ', tagInfo.prop, tagInfo.id);
+	// 	}
+	// };
 
 	const deleteAllFilters = () => {
 		setFormState({
@@ -524,7 +517,10 @@ export const Search: FunctionComponent<SearchProps> = ({ history, location }: Se
 				key={`search-result-${result.id}`}
 				type={result.administrative_type}
 				date={formatDate(result.dcterms_issued)}
-				tags={['Redactiekeuze', 'Partner']}
+				tags={[
+					{ label: 'Redactiekeuze', id: 'redactiekeuze' },
+					{ label: 'Partner', id: 'partner' },
+				]}
 				viewCount={412}
 				bookmarkCount={85}
 				// duration={formatDuration(result.duration_seconds || 0)}
@@ -681,10 +677,12 @@ export const Search: FunctionComponent<SearchProps> = ({ history, location }: Se
 									<FormGroup inlineMode="shrink">
 										<Button label="Zoeken" type="primary" onClick={copySearchTermsToFormState} />
 									</FormGroup>
+									<FormGroup inlineMode="shrink">
+										<Button label="Verwijder alle filters" type="link" onClick={deleteAllFilters} />
+									</FormGroup>
 								</Form>
 							</div>
 						</Spacer>
-						{renderSelectedFilters()}
 						{renderFilterControls()}
 					</Spacer>
 				</Container>
